backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2
django-cors-headers==3.14.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-template-maths==0.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
html5lib==1.1
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
itypes==1.2.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
keyring==23.13.1
lxml==4.9.2
MarkupSafe==2.1.2
more-itertools==9.1.0
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.1
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
pypdf==3.8.1
pypng==0.20220715.0
pyTelegramBotAPI==4.11.0
python-bidi==0.4.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.29.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.7
SecretStorage==3.3.3
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.15
webencodings==0.5.1
xhtml2pdf==0.2.10
zipp==3.15.0
